Move tooltip to its own module
